pyproject.toml
[project]
name = "3omba"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "pyserial>=3.5",
]
